+++ ENETCareData/InterventionGateway.cs
@@ -13,7 +13,7 @@ namespace ENETCareData
         string connectionString = "";
         DatabaseConfig aDatabaseConfig = new DatabaseConfig();
 
-        public int AddNewIntervention(int interventionTypeID, int clientID, double labourRequired, double costRequired, int userID, string interventionDate, string interventionState)
+        public int AddNewIntervention(int interventionTypeID, int clientID, float labourRequired, float costRequired, int userID, string interventionDate, string interventionState)
         {
             int result = 0;
             connectionString = aDatabaseConfig.Setup();
